minor cleanup for PR
Adjusted the debug statements (and removed now-irrelevant files) to make behavior easier to understand for any potential reviewers before merging into OW-994.
