+++ ow_power_system/CMakeLists.txt
@@ -136,7 +136,7 @@ include_directories(
 ## With catkin_make all packages are built within a single CMake context
 ## The recommended prefix ensures that target names across packages don't collide
 add_executable(power_system_node
-  src/main.cpp
+  src/PowerSystemPack.cpp
   src/PowerSystemNode.cpp)
 
 ## Rename C++ executable without prefix
