+++ ow_power_system/include/PowerSystemPack.h
@@ -1,5 +1,5 @@
-#ifndef __POWER_SYSTEM_NODE_H__
-#define __POWER_SYSTEM_NODE_H__
+#ifndef __POWER_SYSTEM_PACK_H__
+#define __POWER_SYSTEM_PACK_H__
 
 #include <vector>
 #include <map>
@@ -8,47 +8,28 @@
 #include <ros/ros.h>
 #include <sensor_msgs/JointState.h>
 #include <PrognoserFactory.h>
+#include "PowerSystemNode.h"
+
+#include "Messages/MessageBus.h"
+#include "Messages/ProgEventMessage.h"
+#include "Messages/ScalarMessage.h"
+#include "ModelBasedAsyncPrognoserBuilder.h"
 
 using PrognoserMap = std::map<PCOE::MessageId, PCOE::Datum<double>>;
 using PrognoserVector = std::vector<PrognoserMap>;
 
-class PowerSystemNode
+const int NUM_NODES = 8;
+
+class PowerSystemPack
 {
 public:
-  PowerSystemNode();
-  bool Initialize(int nodes);
-  void Run();
-  void RunOnce(); // TEST
-  void GetPowerStats(double stored_values[]); // TEST
+  PowerSystemPack();
+  void InitAndRun();
 private:
-  bool loadSystemConfig();
-  PrognoserVector loadPowerProfile(const std::string& filename, std::string custom_file);
-  bool loadCustomFaultPowerProfile(std::string path, std::string custom_file);
-  bool initPrognoser();
+  bool initNodes();
   bool initTopics();
   void jointStatesCb(const sensor_msgs::JointStateConstPtr& msg);
-  double generateTemperatureEstimate();
-  double generateVoltageEstimate();
-  void injectFault(const std::string& power_fault_name,
-                   bool& fault_activated,
-                   double& power,
-                   double& voltage,
-                   double& temperature);
-  void injectCustomFault(bool& fault_activated,
-                         const PrognoserVector& sequence,
-                         size_t& index,
-                         double& wattage,
-                         double& voltage,
-                         double& temperature);
-  void injectFaults(double& power, double& temperature, double& voltage);
-  PrognoserMap composePrognoserData(double power,
-                                    double voltage,
-                                    double temperature);
-  void parseEoD_Event(const ProgEvent& eod_event,
-                      std_msgs::Float64& soc_msg,
-                      std_msgs::Int16& rul_msg,
-                      std_msgs::Float64& battery_temperature_msg);
-  void runPrognoser(double electrical_power);
+  void publishPredictions();
 
   ros::NodeHandle m_nh;                        // Node Handle Initialization
   ros::Publisher m_mechanical_power_raw_pub;   // Mechanical Power Raw
@@ -58,29 +39,29 @@ private:
   ros::Publisher m_battery_temperature_pub;    // Battery Temperature Publisher
   ros::Subscriber m_joint_states_sub;          // Mechanical Power Subscriber
 
+  // The number of PowerSystemNodes to initialize.
+  // WARNING: Code is currently hard-coded based on having 8 PowerSystemNodes.
+  // Changing this value will cause problems.
+  int m_num_nodes = NUM_NODES;
+  std::string node_names[NUM_NODES];
+  PowerSystemNode nodes[NUM_NODES];
+
+  MessageBus bus[NUM_NODES];
+
+  // Store the timestamp/wattage/voltage/temperature of each PowerSystemNode.
+  double models[NUM_NODES][4];
+  double previous_times[NUM_NODES];
+
   int m_moving_average_window = 25;
   std::vector<double> m_power_values;
   size_t m_power_values_index = 0;
 
-  std::unique_ptr<PCOE::Prognoser> m_prognoser;  // Prognoser initialization
-
   std::chrono::time_point<std::chrono::system_clock> m_init_time;
 
-  // Main system configuration: these values are overriden by values
-  // in ../config/system.cfg.
-
-  double m_initial_power = 0.0;         // This is probably always zero
-  double m_initial_temperature = 20.0;  // 20.0 deg. C
-  double m_initial_voltage = 4.1;       // Volts
-  double m_min_temperature = 17.5;      // minimum temp = 17.5 deg. C
-  double m_max_temperature = 21.5;      // maximum temp = 21.5 deg. C
-  double m_battery_lifetime = 2738.0;   // Estimate of battery lifetime (seconds)
-  double m_base_voltage = 3.2;          // [V] estimate
-  double m_voltage_range = 0.1;         // [V]
-  double m_efficiency = 0.9;            // default 90% efficiency
-  double m_gsap_rate_hz = 0.5;          // GSAP's cycle time
-
   double m_current_timestamp = 0.0;
+
+  // GSAP's cycle time
+  double m_gsap_rate_hz = 0.5;
 
   // Baseline value for power drawn by continuously-running systems.
   // This initial value is overriden by the system config.
@@ -101,17 +82,6 @@ private:
   // note that the default values are not expected to change as of
   // Release 9.  This initial value is overriden by the system config.
   int m_profile_increment = 2;
-
-  // The three main statistics relating to battery model health:
-  // Power, Voltage, & Temperature.
-  // Used by the main function to get the health of all models.
-  double m_wattage_estimate;
-  double m_voltage_estimate;
-  double m_temperature_estimate;
-
-  // The number of parallel PowerSystemNodes created. Used for
-  // fault injection modification.
-  int m_total_nodes;
 
   // End main system configuration.
